(PC-14857) refacto: move /home/pages hooks to /home/api directory
