+++ src/features/home/components/ExclusivityModule.tsx
@@ -3,9 +3,9 @@ import { PixelRatio } from 'react-native'
 import FastImage from 'react-native-fast-image'
 import styled from 'styled-components/native'
 
+import { useExcluOffer } from 'features/home/api/useExcluOffer'
 import { shouldDisplayExcluOffer } from 'features/home/components/ExclusivityModule.utils'
 import { ExclusivityPane } from 'features/home/contentful'
-import { useExcluOffer } from 'features/home/pages/useExcluOffer'
 import { useMaxPrice } from 'features/search/utils/useMaxPrice'
 import { analytics } from 'libs/analytics'
 import { useGeolocation } from 'libs/geolocation'
